backend/requirements.txt
google-cloud-alloydb
flask
requests
sqlalchemy
google-cloud-language
"google-cloud-alloydb-connector[asyncpg]"
"cloud-sql-python-connector[pg8000]"